Renamed/moved/cleaned up test classes
