+++ src/main/java/gov/nci/WebAnalytics/PageLoad.java
@@ -1,4 +1,4 @@
-package com.nci.testcases;
+package gov.nci.WebAnalytics;
 
 import java.net.MalformedURLException;
 import java.net.URI;
@@ -7,10 +7,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 import com.nci.Utilities.BrowserManager;
-import gov.nci.WebAnalytics.AnalyticsClick;
-import gov.nci.WebAnalytics.AnalyticsLoad;
-import gov.nci.WebAnalytics.AnalyticsBase;
-
+import com.nci.testcases.AnalyticsTest;
 import com.relevantcodes.extentreports.LogStatus;
 import net.lightbody.bmp.proxy.CaptureType;
 
@@ -21,7 +18,7 @@ import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Parameters;
 import org.testng.annotations.Test;
 
-public class PageLoad extends AnalyticsTestBase {
+public class PageLoad extends AnalyticsTest {
